create @cosmiconfig/core package
