+++ .gitignore
@@ -2,4 +2,4 @@ node_modules
 coverage
 dist
 package-lock.json
-tsconfig.tsbuildinfo
+*.tsbuildinfo
