+++ packages/core/src/merge.ts
