+++ packages/cosmiconfig/test/caches.test.ts
