+++ packages/cosmiconfig/test/failed-directories.test.ts
