+++ packages/cosmiconfig/test/failed-files.test.ts
