+++ packages/cosmiconfig/test/getPropertyByPath.test.ts
