+++ packages/cosmiconfig/test/index.test.ts
