+++ packages/cosmiconfig/test/meta-config.test.ts
