+++ packages/cosmiconfig/test/search-strategies.test.ts
