+++ packages/cosmiconfig/test/successful-directories.test.ts
