+++ packages/cosmiconfig/test/successful-files.test.ts
