+++ packages/cosmiconfig/test/util.ts
