+++ vite.config.ts
@@ -6,11 +6,12 @@ const vitestConfig = defineConfig({
     environment: 'node',
     restoreMocks: true,
     mockReset: true,
-    includeSource: ['src/**/*.{js,ts}'],
+    include: ['packages/*/test/**/*.{test,spec}.?(c|m)[jt]s?(x)'],
+    includeSource: ['packages/*/src/**/*.{js,ts}'],
     coverage: {
       provider: 'istanbul',
       reporter: ['text', 'html', 'lcov'],
-      include: ['src/**/*.{js,ts}'],
+      include: ['packages/*/src/**/*.{js,ts}'],
       branches: 100,
       functions: 100,
       lines: 100,
